refactor: Update DeleteProjectModal to use react-query for check password and delete project mutations
