+++ client/src/components/projects/DeleteProjectModal.tsx
@@ -4,9 +4,10 @@ import { useNavigate, useLocation } from "react-router-dom";
 import { useForm } from "react-hook-form";
 import ErrorMessage from "../ErrorMessage";
 import { CheckPasswordForm } from "@/types/index";
-import { useMutation } from "@tanstack/react-query";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { checkPassword } from "@/api/AuthAPI";
 import { toast } from "react-toastify";
+import { deleteProject } from "@/api/ProjectAPI";
 
 export default function DeleteProjectModal() {
     const initialValues: CheckPasswordForm = {
@@ -14,7 +15,7 @@ export default function DeleteProjectModal() {
     };
     const location = useLocation();
     const navigate = useNavigate();
-
+    const queryClient = useQueryClient();
     const queryParams = new URLSearchParams(location.search);
     const deleteProjectId = queryParams.get("deleteProject")!;
     const show = deleteProjectId ? true : false;
@@ -24,18 +25,29 @@ export default function DeleteProjectModal() {
         handleSubmit,
         formState: { errors },
     } = useForm({ defaultValues: initialValues });
+
     const checkUserPasswordMutation = useMutation({
         mutationFn: checkPassword,
         onError: (error) => {
             toast.error(error.message);
         },
+    });
+
+    const deleteProjectMutation = useMutation({
+        mutationFn: deleteProject,
+        onError: (error) => {
+            toast.error(error.message);
+        },
         onSuccess: (data) => {
+            queryClient.invalidateQueries({ queryKey: ["projects"] });
             toast.success(data);
+            navigate(location.pathname, { replace: true });
         },
     });
+
     const handleForm = async (formData: CheckPasswordForm) => {
         await checkUserPasswordMutation.mutateAsync(formData);
-        console.log(`se muto parece`);
+        await deleteProjectMutation.mutateAsync(deleteProjectId);
     };
 
     return (
