+++ client/src/views/DashboardView.tsx
@@ -2,9 +2,8 @@ import { Fragment } from "react";
 import { Menu, Transition } from "@headlessui/react";
 import { EllipsisVerticalIcon } from "@heroicons/react/20/solid";
 import { Link, useLocation, useNavigate } from "react-router-dom";
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import { deleteProject, getProjects } from "@/api/ProjectAPI";
-import { toast } from "react-toastify";
+import { useQuery } from "@tanstack/react-query";
+import { getProjects } from "@/api/ProjectAPI";
 import { useAuth } from "@/hooks/useAuth";
 import { isManager } from "@/utils/polices";
 import DeleteProjectModal from "@/components/projects/DeleteProjectModal";
@@ -17,17 +16,7 @@ export default function DashboardView() {
         queryKey: ["projects"],
         queryFn: getProjects,
     });
-    const queryClient = useQueryClient();
-    const { mutate } = useMutation({
-        mutationFn: deleteProject,
-        onError: (error) => {
-            toast.error(error.message);
-        },
-        onSuccess: (data) => {
-            queryClient.invalidateQueries({ queryKey: ["projects"] });
-            toast.success(data);
-        },
-    });
+
     if (isLoading && authLoading) return "Cargando..";
     if (data && user)
         return (
@@ -135,7 +124,10 @@ export default function DashboardView() {
                                                                 type="button"
                                                                 className="block px-3 py-1 text-sm leading-6 text-red-500"
                                                                 onClick={() =>
-                                                                    navigate(location.pathname + `?deleteProject=${project._id}`)
+                                                                    navigate(
+                                                                        location.pathname +
+                                                                            `?deleteProject=${project._id}`
+                                                                    )
                                                                 }
                                                             >
                                                                 Eliminar
